fix: added docker and tests

+++ internal/client/user/grpc/grpc.go
@@ -22,12 +22,12 @@ func New(
 	ctx context.Context,
 	addr string,
 	timeout time.Duration,
-	retriesCount int,
+	retriesCount uint,
 ) (*Client, error) {
 
 	retryOpts := []retry.CallOption{
 		retry.WithCodes(codes.NotFound, codes.Aborted, codes.DeadlineExceeded),
-		retry.WithMax(uint(retriesCount)),
+		retry.WithMax(retriesCount),
 		retry.WithPerRetryTimeout(timeout),
 	}
